Add generic namespace for handling different types of subscriptions.
The namespace is responsible for translating between arbitrary
incoming messages for a given SocketIO endpoint and a specific iex message
type, such as TOPS. It is also responsible for forwarding the decoded
messages to all concerned listeners. Lastly, it filters out messages that
given listeners are not concerned with. All this allows multiplexing of
endpoint subscriptions over the same SocketIO connection.

Also simplifies encoding to handle only messages of known formats rather than
trying to handle arbitrary structs in a consistent manner.

Adds a subUnsubMsgFactory function type for generating subscribe and unsubscribe
messages for symbols across the given endpoint types.